requirements.txt
python-docx>=1.1.0
jinja2>=3.1.0
python-frontmatter>=1.1.0
pydantic>=2.6.0
typer>=0.12.0
openpyxl>=3.1.0
Pillow>=10.0.0
streamlit>=1.30.0
pandas>=2.0.0
